[MKLevinController] immediately scan directories which are added to the userdefaults
The arraycontroller managing the userdefaults array
of paths to scan is observed for changes in the
content array.
Due to a bug in NSArrayController it is not possible
to recieve any other value than NSKeyValueChangeSetting
for the NSKeyValueChangeKindKey entry in the change
dictionary.
Therefore a copy of the content array is created
everytime something gets updated and the old
size is compared to the new size to check whether
something has been added or removed.